Added Help commands
Modified Insult to take a parameter.  Check to see if that person is a bot.
Prevented the entire bot from crashing if a wiki write hits an error.  Poops out an exception in irc.
Fixed an issue with the script_help2 parser where the description of the function wasn't grabbed.